refactor(backend): move desktop template builder into SDK module
